renaming TreeLines so that the change of casing can be committed
