+++ lib/index.js
@@ -99,7 +99,7 @@ export { Windowing } from './Windowing';
 export FinderPanelLayout from './FinderPanelLayout';
 export { PillInput } from './PillInput/PillInput/PillInput';
 export DismissiblePrompt from './DismissiblePrompt';
-export TreeLines from './TreeLines';
+export TreeLines from './TreeLinesss';
 export {
   ButtonPrimary
 } from './src/modules/Button/ButtonPrimary/ButtonPrimary';
