+++ styles/components/_ui.scss
@@ -61,7 +61,7 @@
 @import '../../lib/SelectionModal/styles.scss';
 @import '../../lib/InputConfirmationModal/styles.scss';
 @import '../../lib/Layout/styles.scss';
-@import '../../lib/Treelines/styles.scss';
+@import '../../lib/Treelinesss/styles.scss';
 
 @import '../../lib/src/modules/Button/styles/styles.scss';
